ai/requirements.txt
Flask==3.0.3
urllib3==1.26.15
python-dotenv==1.0.1
langchain==0.3.7
langchain_core==0.3.19
langchain_pinecone==0.2.0
langchain_huggingface==0.1.2
langchain_community==0.3.7
langchain_openai==0.2.8
flask_cors==5.0.0
python-dotenv==1.0.1
